requirements.txt
# File: requirements.txt

requests>=2.28.0
python-dotenv>=0.19.0
PyYAML>=6.0
urllib3>=2.0.0
typing-extensions>=4.0.0
python-dateutil>=2.8.0
fcntl-linux  # For file locking on Linux
pytz        # For timezone handling 
pyyaml      # For config file parsing